% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coreNetAnalysis.R
\name{coreNetAnalysis}
\alias{coreNetAnalysis}
\title{Run Graphical Analysis Core Parameters}
\usage{
coreNetAnalysis(ginp)
}
\arguments{
\item{ginp}{The prepared graph object from prepareGraphs function}
}
\value{
Gives the edge and weighted edge counts, number of nodes, density, degree (averages), memberships, modularity, centrality, articulation points, and strong/weak plots as a list object
}
\description{
Analyzes the graphs with the core parameters, such as number of edges and nodes, density, average degree, centrality, and modularity
}
\examples{
df <- sampleData1
prepNet <- tabulate_edges(df, iscsvfile = FALSE, silentNodes = 0)
baseNet <- prepareGraphs(prepNet, project_title = "Sample Data 1", weightedGraph = TRUE)

coreNetAnalysis(baseNet)

}
