% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgelist_raw.R
\name{edgelist_raw}
\alias{edgelist_raw}
\title{Process raw order lists from two column format to edge lists}
\usage{
edgelist_raw(input_file)
}
\arguments{
\item{input_file}{Source of the raw input file. Must be in a .csv format}
}
\value{
Saves the weight and edge lists as a data.frame object or a .csv file to disk.
}
\description{
Takes raw input that is in a 2 column format/question-and-response format and generates an appropriate edge lists in a combined .csv file.
}
\examples{
df <- sampleData1
prepNet <- edgelist_raw(df)

}
