% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleData1.R
\docType{data}
\name{sampleData1}
\alias{sampleData1}
\title{Sample Episode Start and Episode Continuation Data (2 Column)}
\format{
A data frame with 466 rows and 2 variables:
\describe{
\item{ep_start}{If participant in the graph has initiated an episode}
\item{ep_cont}{If participant in the graph has raised a response to the previous episode}
...
}
}
\usage{
sampleData1
}
\description{
A data set that contains Episode Start (in this case, questions) and Episode Contination (in this case, responses) data in a
2 column format. This data was randomly generated. Case basis: 4 students in a group discussion questions in a STEM course
at a major university.
}
\keyword{datasets}
